+++ pkg/server/api/authn/cookie.go
@@ -34,6 +34,10 @@ func (c *Cookie) Set(w http.ResponseWriter, session *coredata.Session) {
 	)
 }
 
+func (c *Cookie) Clear(w http.ResponseWriter, session *coredata.Session) {
+	securecookie.Clear(w, c.sessionCookieConfig(time.Until(session.ExpiredAt)))
+}
+
 func NewCookie(config *securecookie.Config) *Cookie {
 	return &Cookie{config}
 }